refactor: move type cast and assertion utils into their own file
