+++ src/rules/utils.ts
@@ -16,41 +16,6 @@ export const createRule = ESLintUtils.RuleCreator(name => {
   return `${REPO_URL}/blob/v${version}/docs/rules/${ruleName}.md`;
 });
 
-export type MaybeTypeCast<Expression extends TSESTree.Expression> =
-  | TSTypeCastExpression<Expression>
-  | Expression;
-
-type TSTypeCastExpression<
-  Expression extends TSESTree.Expression = TSESTree.Expression,
-> = AsExpressionChain<Expression> | TypeAssertionChain<Expression>;
-
-interface AsExpressionChain<
-  Expression extends TSESTree.Expression = TSESTree.Expression,
-> extends TSESTree.TSAsExpression {
-  expression: AsExpressionChain<Expression> | Expression;
-}
-
-interface TypeAssertionChain<
-  Expression extends TSESTree.Expression = TSESTree.Expression,
-> extends TSESTree.TSTypeAssertion {
-  expression: TypeAssertionChain<Expression> | Expression;
-}
-
-const isTypeCastExpression = <Expression extends TSESTree.Expression>(
-  node: MaybeTypeCast<Expression>,
-): node is TSTypeCastExpression<Expression> =>
-  node.type === AST_NODE_TYPES.TSAsExpression ||
-  node.type === AST_NODE_TYPES.TSTypeAssertion;
-
-export const followTypeAssertionChain = <
-  Expression extends TSESTree.Expression,
->(
-  expression: MaybeTypeCast<Expression>,
-): Expression =>
-  isTypeCastExpression(expression)
-    ? followTypeAssertionChain(expression.expression)
-    : expression;
-
 /**
  * A `Literal` with a `value` of type `string`.
  */
@@ -339,3 +304,4 @@ export const getTestCallExpressionsFromDeclaredVariables = (
 export * from './utils/parseJestFnCall';
 export * from './utils/detectJestVersion';
 export * from './utils/parseExpectCall';
+export * from './utils/followTypeAssertionChain';
